initialized trueMFs to true, else all mfnc synapse drop to zero after the first trial. tested mfnc and pfpc weights saving and loading, refactored spike_sum calculation code

+++ src/cbm_core/cbmsimcore.h
@@ -72,7 +72,7 @@ private:
 	int numGRStim    =  0;
 	int startGRStim  =  0;
 
-	unsigned long curTime;
+	uint32_t curTime;
 
 	void construct(CBMState *state, int *mzoneRSeed,
 		int gpuIndStart, int numGPUP2);
